requirements.txt
# SVN自动合并智能体工具依赖
requests>=2.28.0
schedule>=1.2.0
colorama>=0.4.6
rich>=13.0.0
ollama>=0.1.0
python-dotenv>=1.0.0
